+++ src/app/clusters/commissioner-control-server/CodegenIntegration.h
@@ -1,6 +1,6 @@
 /*
- *    Copyright (c) 2025-2026 Project CHIP Authors
- *    All rights reserved.
+ *
+ *    Copyright (c) 2025 Project CHIP Authors
  *
  *    Licensed under the Apache License, Version 2.0 (the "License");
  *    you may not use this file except in compliance with the License.
@@ -14,6 +14,7 @@
  *    See the License for the specific language governing permissions and
  *    limitations under the License.
  */
+
 #pragma once
 
-#include <app/clusters/commissioner-control-server/CodegenIntegration.h>
+#include <app/clusters/commissioner-control-server/CommissionerControlCluster.h>
