+++ src/app/clusters/commissioner-control-server/CommissionerControlCluster.h
@@ -17,89 +17,10 @@
  */
 #pragma once
 
-#include <app-common/zap-generated/cluster-objects.h>
 #include <app/CommandHandlerInterface.h>
+#include <app/clusters/commissioner-control-server/Delegate.h>
 
-namespace chip {
-namespace app {
-namespace Clusters {
-namespace CommissionerControl {
-
-// Spec indicates that IP Address is either 4 or 16 bytes.
-static constexpr size_t kIpAddressBufferSize = 16;
-
-struct CommissioningApprovalRequest
-{
-    uint64_t requestId;
-    VendorId vendorId;
-    uint16_t productId;
-    NodeId clientNodeId;
-    FabricIndex fabricIndex;
-    Optional<CharSpan> label;
-};
-
-struct CommissioningWindowParams
-{
-    uint32_t iterations;
-    uint16_t commissioningTimeout;
-    uint16_t discriminator;
-    ByteSpan PAKEPasscodeVerifier;
-    ByteSpan salt;
-};
-
-class Delegate
-{
-public:
-    /**
-     * @brief Handle a commissioning approval request.
-     *
-     * This command is sent by a client to request approval for a future CommissionNode call.
-     * The server SHALL always return SUCCESS to a correctly formatted RequestCommissioningApproval
-     * command, and then send a CommissioningRequestResult event once the result is ready.
-     *
-     * @param request The commissioning approval request to handle.
-     * @return CHIP_ERROR indicating the success or failure of the operation.
-     */
-    virtual CHIP_ERROR HandleCommissioningApprovalRequest(const CommissioningApprovalRequest & request) = 0;
-
-    /**
-     * @brief Validate a commission node command.
-     *
-     * The server SHALL return FAILURE if the CommissionNode command is not sent from the same
-     * NodeId as the RequestCommissioningApproval or if the provided RequestId to CommissionNode
-     * does not match the value provided to RequestCommissioningApproval.
-     *
-     * The validation SHALL fail if the client Node ID is kUndefinedNodeId, such as getting the NodeID from
-     * a group or PASE session.
-     *
-     * @param clientNodeId The NodeId of the client.
-     * @param requestId The request ID to validate.
-     * @return CHIP_ERROR indicating the success or failure of the operation.
-     */
-    virtual CHIP_ERROR ValidateCommissionNodeCommand(NodeId clientNodeId, uint64_t requestId) = 0;
-
-    /**
-     * @brief Get the parameters for the commissioning window.
-     *
-     * This method is called to retrieve the parameters needed for the commissioning window.
-     *
-     * @param[out] outParams The parameters for the commissioning window.
-     * @return CHIP_ERROR indicating the success or failure of the operation.
-     */
-    virtual CHIP_ERROR GetCommissioningWindowParams(CommissioningWindowParams & outParams) = 0;
-
-    /**
-     * @brief Handle a commission node request.
-     *
-     * Commission a node specified by the previously approved request.
-     *
-     * @param params The parameters for the commissioning window.
-     * @return CHIP_ERROR indicating the success or failure of the operation.
-     */
-    virtual CHIP_ERROR HandleCommissionNode(const CommissioningWindowParams & params) = 0;
-
-    virtual ~Delegate() = default;
-};
+namespace chip::app::Clusters {
 
 class CommissionerControlServer : public CommandHandlerInterface
 {
@@ -111,7 +32,7 @@ public:
      * Note: the caller must ensure that the delegate lives throughout the instance's lifetime.
      * @param endpointId The endpoint on which this cluster exists. This must match the zap configuration.
      */
-    CommissionerControlServer(Delegate * delegate, EndpointId endpointId);
+    CommissionerControlServer(CommissionerControl::Delegate * delegate, EndpointId endpointId);
 
     ~CommissionerControlServer() override;
 
@@ -122,12 +43,12 @@ public:
      */
     CHIP_ERROR Init();
 
-    Protocols::InteractionModel::Status
-    GetSupportedDeviceCategoriesValue(EndpointId endpoint,
-                                      BitMask<SupportedDeviceCategoryBitmap> * supportedDeviceCategories) const;
+    Protocols::InteractionModel::Status GetSupportedDeviceCategoriesValue(
+        EndpointId endpoint, BitMask<CommissionerControl::SupportedDeviceCategoryBitmap> * supportedDeviceCategories) const;
 
     Protocols::InteractionModel::Status
-    SetSupportedDeviceCategoriesValue(EndpointId endpoint, const BitMask<SupportedDeviceCategoryBitmap> supportedDeviceCategories);
+    SetSupportedDeviceCategoriesValue(EndpointId endpoint,
+                                      const BitMask<CommissionerControl::SupportedDeviceCategoryBitmap> supportedDeviceCategories);
 
     /**
      * @brief
@@ -149,7 +70,7 @@ private:
      * If the operational state is not in 'Stopped', returns the Interaction Model status code of INVALID_IN_STATE.
      */
     void HandleRequestCommissioningApproval(HandlerContext & ctx,
-                                            const Commands::RequestCommissioningApproval::DecodableType & req);
+                                            const CommissionerControl::Commands::RequestCommissioningApproval::DecodableType & req);
 
     /**
      * @brief Handle Command: AddMoreTime.
@@ -157,12 +78,11 @@ private:
      * If the cook time value is out of range, returns the Interaction Model status code of CONSTRAINT_ERROR.
      * If the operational state is in 'Error', returns the Interaction Model status code of INVALID_IN_STATE.
      */
-    void HandleCommissionNode(HandlerContext & ctx, const Commands::CommissionNode::DecodableType & req);
+    void HandleCommissionNode(HandlerContext & ctx, const CommissionerControl::Commands::CommissionNode::DecodableType & req);
 
-    Delegate * mDelegate = nullptr;
+    CommissionerControl::Delegate * mDelegate = nullptr;
 };
 
-} // namespace CommissionerControl
 } // namespace Clusters
 } // namespace app
 } // namespace chip
